Use intptr_t for loop bounds in OpenMP implementation

+++ src/omp/OMPStream.cpp
@@ -13,7 +13,7 @@
 #endif
 
 template <class T>
-OMPStream<T>::OMPStream(const long long ARRAY_SIZE, int device)
+OMPStream<T>::OMPStream(const intptr_t ARRAY_SIZE, int device)
 {
   array_size = ARRAY_SIZE;
 
@@ -39,7 +39,7 @@ OMPStream<T>::~OMPStream()
 {
 #ifdef OMP_TARGET_GPU
   // End data region on device
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *a = this->a;
   T *b = this->b;
   T *c = this->c;
@@ -54,7 +54,7 @@ OMPStream<T>::~OMPStream()
 template <class T>
 void OMPStream<T>::init_arrays(T initA, T initB, T initC)
 {
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
 #ifdef OMP_TARGET_GPU
   T *a = this->a;
   T *b = this->b;
@@ -63,7 +63,7 @@ void OMPStream<T>::init_arrays(T initA, T initB, T initC)
 #else
   #pragma omp parallel for
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     a[i] = initA;
     b[i] = initB;
@@ -89,7 +89,7 @@ void OMPStream<T>::read_arrays(std::vector<T>& h_a, std::vector<T>& h_b, std::ve
 #endif
 
   #pragma omp parallel for
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     h_a[i] = a[i];
     h_b[i] = b[i];
@@ -102,14 +102,14 @@ template <class T>
 void OMPStream<T>::copy()
 {
 #ifdef OMP_TARGET_GPU
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *a = this->a;
   T *c = this->c;
   #pragma omp target teams distribute parallel for simd
 #else
   #pragma omp parallel for
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     c[i] = a[i];
   }
@@ -126,14 +126,14 @@ void OMPStream<T>::mul()
   const T scalar = startScalar;
 
 #ifdef OMP_TARGET_GPU
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *b = this->b;
   T *c = this->c;
   #pragma omp target teams distribute parallel for simd
 #else
   #pragma omp parallel for
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     b[i] = scalar * c[i];
   }
@@ -148,7 +148,7 @@ template <class T>
 void OMPStream<T>::add()
 {
 #ifdef OMP_TARGET_GPU
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *a = this->a;
   T *b = this->b;
   T *c = this->c;
@@ -156,7 +156,7 @@ void OMPStream<T>::add()
 #else
   #pragma omp parallel for
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     c[i] = a[i] + b[i];
   }
@@ -173,7 +173,7 @@ void OMPStream<T>::triad()
   const T scalar = startScalar;
 
 #ifdef OMP_TARGET_GPU
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *a = this->a;
   T *b = this->b;
   T *c = this->c;
@@ -181,7 +181,7 @@ void OMPStream<T>::triad()
 #else
   #pragma omp parallel for
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     a[i] = b[i] + scalar * c[i];
   }
@@ -198,7 +198,7 @@ void OMPStream<T>::nstream()
   const T scalar = startScalar;
 
 #ifdef OMP_TARGET_GPU
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *a = this->a;
   T *b = this->b;
   T *c = this->c;
@@ -206,7 +206,7 @@ void OMPStream<T>::nstream()
 #else
   #pragma omp parallel for
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     a[i] += b[i] + scalar * c[i];
   }
@@ -223,14 +223,14 @@ T OMPStream<T>::dot()
   T sum = 0.0;
 
 #ifdef OMP_TARGET_GPU
-  int array_size = this->array_size;
+  intptr_t array_size = this->array_size;
   T *a = this->a;
   T *b = this->b;
   #pragma omp target teams distribute parallel for simd map(tofrom: sum) reduction(+:sum)
 #else
   #pragma omp parallel for reduction(+:sum)
 #endif
-  for (long long i = 0; i < array_size; i++)
+  for (intptr_t i = 0; i < array_size; i++)
   {
     sum += a[i] * b[i];
   }
